+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## [Unreleased]
 
+## [0.2.0] - 2023-10-07
+
+- Feature to Pause/Unpause Periodic Jobs
+
 ## [0.1.0] - 2023-10-07
 
-- Run manualy Periodic Jobs
+- Feature to run manualy Periodic Jobs